backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.7
django-channels==0.7.0
django-cors-headers==4.7.0
django-ipware==7.0.1
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
djoser==2.3.1
jwt==1.3.1
pyjwt==1.7.1
msgpack==1.1.0
oauthlib==3.2.2
python-dotenv==1.0.1
python-http-client==3.3.7
python3-openid==3.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
sendgrid==6.11.0
setuptools==65.5.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tzdata==2025.1
urllib3==2.3.0
